+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the "alloglot" extension will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 This project adhere's to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+- Configure path to `grep`.
+
 ## [2.6.0]
 
 - Add `deactivateCommand` to `TConfig` to run on extension `deactivate()`