+++ CHANGELOG.md
@@ -6,7 +6,10 @@ This project adhere's to [Semantic Versioning](https://semver.org/spec/v2.0.0.ht
 
 ## [unreleased]
 
-- Configure path to `grep`.
+- Support configurable path to `grep`.
+- Support multiple tags configs. `alloglot.tags` must now be an array of `TagsConfig`.
+- Support merging workspace config with fallback config.
+- Some error handling and reporting for configs.
 
 ## [2.6.0]