+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to the "alloglot" extension will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 This project adhere's to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+- Support configurable path to `grep`.
+- Support multiple tags configs. `alloglot.tags` must now be an array of `TagsConfig`.
+- Support merging workspace config with fallback config.
+- Some error handling and reporting for configs.
+
 ## [2.6.0]
 
 - Add `deactivateCommand` to `TConfig` to run on extension `deactivate()`